city-dashboard/edge/linux/requirements.txt
awsiotsdk>=1.21
python-dotenv>=1.0
